Add needed const qualifier to return value of StringCopy
Copied strings are burried so deep that it (almost) caused bugs like
writing to read-only memory and crashing, or making a single digit
string copy as something else. Alternative solution would be to drop
this StaticMem_t "static zone memory" allocation, or not use
StringCopy in several places (files.cpp).