Change G2_LISTBONES and G2_LISTSURFACES syscalls
They used to work on x86 because they take an untranslated pointer to
CGhoul2Info as an argument. Modules have no legitimate way of learning
about such pointer now and it's not safe to call these G2 API
functions with potentially wrong address. They are not used by
original modules and the only output/side effect they produce
is printing debug/development text in the console.